Fix discovery of modules in namespace packages
Instead of relying on `__init__.py` files, stop at the first parent
directory that is in `sys.path`. This gives the shortest module name
under which the file can really be imported. (Unless there are name
conflicts in `sys.path`, which is arguably a misconfiguration; this
is caught by the location check in `module_tree`.)